pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ligand-neff"
version = "0.2.1"
description = "Multiple Ligand Alignment and Neff computation"
authors = [
    { name = "Author" }
]
requires-python = ">=3.10"
dependencies = [
    "rdkit",
    "jax==0.5.0",
    "jaxlib",
    "numpy>=1.24",
    "chex>=0.1.87",
    "jaxtyping>=0.2.28",
    "beartype>=0.18",
    "pyyaml",
    "pytest>=9.0.2",
    "pandas>=2.3.3",
    "matplotlib>=3.10.8",
    "jax-tqdm>=0.4.0",
]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
